requirements.txt
anyio==3.6.1
certifi==2022.6.15
click==8.1.3
fastapi==0.72.0
h11==0.13.0
httptools==0.6.1
idna==3.3
pip==21.3.1
pydantic==1.9.1
python-decouple==3.8
PyYAML==6.0
psutil==5.9.8
mongoengine==0.23.1
mongomock==3.19.0
mongoengine-goodjson==1.1.8
pymongo==3.12.0
sentry-sdk==1.8.0
setuptools==60.2.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.19.0
watchfiles==0.16.1
websockets==10.3
wheel==0.37.1